Fix something that jinja2>=3 made an error
In jinja2==2.* `super in typedefs` used to evaluate to false if the
third argument to the macro (typedefs) wasn’t passed.

In jinja2==3.*, this becomes an error instead (which seems right and
proper to me). The easiest solution is to explicitly show that typedefs
is optional by specifying a default value.